Improve FactCategory collectionView
