+++ Chuck Norris Facts/Scenes/Facts/SearchFacts/Cells/FactCategoryViewFlowLayout.swift
@@ -1,0 +1,32 @@
+//
+//  FactCategoryViewFlowLayout.swift
+//  Chuck Norris Facts
+//
+//  Created by Djorkaeff Alexandre Vilela Pereira on 10/21/20.
+//  Copyright © 2020 Djorkaeff Alexandre Vilela Pereira. All rights reserved.
+//
+
+import UIKit
+
+final class FactCategoryViewFlowLayout: UICollectionViewFlowLayout {
+
+    override func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]? {
+        let attributes = super.layoutAttributesForElements(in: rect)
+
+        var leftMargin = sectionInset.left
+        var maxY: CGFloat = -1.0
+        attributes?.forEach { layoutAttribute in
+            if layoutAttribute.representedElementCategory == .cell {
+                if layoutAttribute.frame.origin.y >= maxY {
+                    leftMargin = sectionInset.left
+                }
+                layoutAttribute.frame.origin.x = leftMargin
+                leftMargin += layoutAttribute.frame.width + minimumInteritemSpacing
+                maxY = max(layoutAttribute.frame.maxY, maxY)
+            }
+        }
+
+        return attributes
+    }
+
+}
